Fixed some bugs (combat logging interfering with mini-games, water things in WorldEater) and introduced user data removal feature (though not implemented fully yet).

+++ src/main/java/org/noxet/noxetserver/combatlogging/CombatLogging.java
@@ -8,6 +8,7 @@ import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.scheduler.BukkitTask;
 import org.noxet.noxetserver.NoxetServer;
 import org.noxet.noxetserver.messaging.Message;
+import org.noxet.noxetserver.minigames.MiniGameManager;
 
 import java.util.*;
 
@@ -18,9 +19,12 @@ public class CombatLogging {
     private static final int combatLogTimeout = 20 * 20;
 
     public static void triggerCombatLog(Player player) {
+        if(MiniGameManager.isPlayerBusyInGame(player))
+            return;
+
         BukkitTask existingTask = combatLogged.remove(player);
         if(existingTask != null)
-            existingTask.cancel(); // If already combat logged: cancel the timeout and set a new one:
+            existingTask.cancel(); // If already combat logged: cancel the timeout and set a new one.
         else
             new Message("§c§lCOMBAT!§7 Logging out or teleporting away while in combat will kill you.").send(player);
 
@@ -66,5 +70,11 @@ public class CombatLogging {
 
         playerInventory.clear();
         new CombatLoggingStorageManager().setPlayerCombatLogMode(player, getCurrentRealm(player), true);
+    }
+
+    public static void clearCombatLog(Player player) {
+        BukkitTask timeoutTask = combatLogged.remove(player);
+        if(timeoutTask != null)
+            timeoutTask.cancel();
     }
 }
